extracting theming to common-ui module
