+++ versions.properties
@@ -7,40 +7,79 @@
 #### suppress inspection "SpellCheckingInspection" for whole file
 #### suppress inspection "UnusedProperty" for whole file
 
-version.androidx.appcompat=1.2.0
+version.androidx.appcompat=1.3.0
+
 version.androidx.arch.core=2.1.0
+
 version.androidx.constraintlayout=2.0.4
+
+version.androidx.localbroadcastmanager=1.0.0
+##                         # available=1.1.0-alpha01
+
 version.androidx.recyclerview=1.2.1
-version.androidx.core=1.3.2
+
+version.androidx.core=1.6.0
+
 version.androidx.fragment=1.2.5
+
 version.androidx.legacy=1.0.0
+
 version.androidx.lifecycle=2.4.0-alpha01
+
 version.androidx.room=2.2.6
+
 version.androidx.swiperefreshlayout=1.1.0
+
 version.androidx.test=1.4.0-beta02
+
 version.androidx.webkit=1.3.0
+
 version.androidx.work=2.4.0
+
 version.androidx.security=1.1.0-alpha03
+
 version.androidx.security-crypto=1.1.0-alpha03
+
 version.com.airbnb.android..lottie=3.7.0
+
 version.com.android.installreferrer..installreferrer=1.1.2
+
 version.com.github.bumptech.glide..compiler=4.12.0
+
 version.com.github.bumptech.glide..glide=4.12.0
+
 version.com.github.bumptech.glide..okhttp3-integration=4.12.0
+
 version.com.jakewharton.rxrelay2..rxrelay=2.0.0
+
 version.com.jakewharton.timber..timber=4.7.1
+
 version.com.nhaarman.mockitokotlin2..mockito-kotlin=2.2.0
+
 version.google.android.material=1.2.1
+
 version.google.dagger=2.37
+
 version.io.reactivex.rxjava2..rxandroid=2.0.2
+
 version.io.reactivex.rxjava2..rxjava=2.2.9
+
 version.kotlin=1.4.10
+
 version.kotlinx.coroutines=1.5.0
+
 version.leakcanary=2.7
+
 version.okhttp3=4.9.0
+
 version.org.apache.commons..commons-math3=3.6.1
+
 version.mockito=3.4.6
+
 version.retrofit2=2.9.0
+
 version.junit.junit=4.13.2
+
 version.org.jsoup..jsoup=1.13.1
+
 version.app.cash.turbine..turbine=0.5.2
